Complete the project URL calculation unification
Complete the changes started in version 4.5 to normalize how the
project/installation title and URL are calculated. The setting
`COMMON_PROJECT_URL` is now removed as its intended purpose is now
performed by the settings `ORGANIZATIONS_INSTALLATION_URL` and
`ORGANIZATIONS_URL_BASE_PATH`. This change also fixes the title of the
REST API documentation showing the text 'None' when the
`COMMON_PROJECT_TITLE` setting was left to its default value.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>