Pass `log_entry` to workflow state actions
Pass the workflow instance log entry to the workflow instance states.

This is to allow states to have access to the log entry that triggered
the workflow transition regardless of the current state or transition.

The workflow state action context is now composed of the keys:
`workflow_instance`, `workflow_instance_context`, `action`, and
`log_entry`.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>