Settings app behavior changes
Add support for ignoring setting formatting errors. This will allow
the system to start even when there are settings with badly formatted
values.

This does no apply to bootstrap settings which are low level do need
to be correct in order for the system to start.

Add the setting `SETTINGS_IGNORE_ERRORS` which defaults to `True`.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>